fix: strange spacing on code blocks
Fixes (#2)

This is a pretty dirty fix, and I would like to figure out a proper way of doing this.

When the content of the code block triggers horizontal scrolling, the overflow doesn't respect the padding value set by the parent pre element.

A way around this is to add margin right to each line by selecting the data-line attribute. We are using 16px to match the padding set on the pre element.